tools/ios-agent/requirements.txt
pymobiledevice3
appium-python-client
Pillow
requests
pyyaml
colorama
